Refactor access duration types across contracts to uint48
- Updated the access duration parameter type from uint64 to uint48 in various contracts including `AppRegistry`, `BaseApp`, `IAppAccount`, and `ExecutorBase` to standardize duration handling.
- Adjusted related functions and structures to reflect the new duration type, ensuring consistency across the codebase.
- Modified test cases to validate the changes in duration handling, ensuring compatibility with the updated types.

+++ packages/contracts/src/apps/BaseApp.sol
@@ -39,7 +39,7 @@ abstract contract BaseApp is ITownsApp {
         return _installPrice();
     }
 
-    function accessDuration() external view returns (uint64) {
+    function accessDuration() external view returns (uint48) {
         return _accessDuration();
     }
 
@@ -52,5 +52,5 @@ abstract contract BaseApp is ITownsApp {
 
     function _installPrice() internal view virtual returns (uint256) {}
 
-    function _accessDuration() internal view virtual returns (uint64) {}
+    function _accessDuration() internal view virtual returns (uint48) {}
 }
