+++ packages/contracts/src/spaces/facets/executor/ExecutorBase.sol
@@ -42,7 +42,7 @@ abstract contract ExecutorBase is IExecutorBase {
     /// @param groupId The ID of the group to create.
     /// @param status The status to set (active/inactive).
     /// @param expiration Optional timestamp when the group should expire (0 for no expiration).
-    function _setGroupStatus(bytes32 groupId, bool status, uint64 expiration) internal {
+    function _setGroupStatus(bytes32 groupId, bool status, uint48 expiration) internal {
         Group storage group = _getGroup(groupId);
         group.setStatus(status);
         if (status && expiration > 0) {
@@ -52,7 +52,7 @@ abstract contract ExecutorBase is IExecutorBase {
         emit GroupStatusSet(groupId, status);
     }
 
-    function _getGroupExpiration(bytes32 groupId) internal view returns (uint64) {
+    function _getGroupExpiration(bytes32 groupId) internal view returns (uint48) {
         return _getGroup(groupId).expiration;
     }
 
@@ -66,7 +66,7 @@ abstract contract ExecutorBase is IExecutorBase {
     /// @notice Sets or extends the expiration for a group.
     /// @param groupId The ID of the group.
     /// @param expiration The new expiration timestamp.
-    function _setGroupExpiration(bytes32 groupId, uint64 expiration) internal {
+    function _setGroupExpiration(bytes32 groupId, uint48 expiration) internal {
         if (expiration <= block.timestamp) revert InvalidExpiration();
         Group storage group = _getGroup(groupId);
         group.expiration = expiration;
